# added real-time / speed up to server request