removed redundant model creation and removed io calls that saved the data server-side 

Stats:
capture request without gpu: ~ 1.6s
capture request with gpu: ~ 1s
getCoordsFast (needs a rename) without gpu = ~0.7s
getCoordsFast with gpu = ~0.4s